+++ doc/source/changelog.rst
@@ -2,6 +2,20 @@
  Changelog
 ==========
 
+v1.4.1
+======
+
+* Implemented the ``wireframe`` option on ``plot3d`` and
+  ``plot3d_parametric_surface`` to add a "nice" wireframe over the surface.
+
+* Fixed bug within ``PlotlyBackend``: it should now be able to plot more than
+  14 2d/3d parametric lines when ``use_cm=False``.
+
+* Fixed bug with ``plot3d`` and ``plot_contour`` when dealing with instances
+  of ``BaseScalar``.
+
+
+
 v1.4.0
 ======